cards v0.1.0 Cards

Provides methods for creating and handling a deck of cards

Functions

contains?(deck, card)

create_deck()

Returns a list of strings representing a deck of playing cards

create_hand(hand_size)

deal(deck, amount_Of_Cards)

Divides a deck into a hand and remainder of the deck. The hand_size argument indicates how many cards should be in the hand.

Examples

iex> deck = Cards.create_deck()
iex> {hand, deck} = Cards.deal(deck, 1)
iex> hand
["Ace of Spades"]

load(filename)

save(deck, filename)

shuffle(deck)